scripts/requirements.txt
azure-identity==1.16.0
azure-search-documents==11.4.0
openai[datalib]==1.30.5
tenacity==8.3.0
redis==5.0.4
postgres==4.0
beautifulsoup4==4.12.3
keybert==0.8.4
psycopg2-binary==2.9.9